Added sample data to enable runs without chedispy library. Refactored 03-generate-sample-data
