+++ 00-README.md
@@ -20,11 +20,9 @@ ___
 | Library Name   | Library License | Library License URL |  Library Source URL |
 |----------------------------------------|-------------------------|------------|-----------------------------------------------------|
 | Apache Spark     | Apache License 2.0 | 	https://github.com/apache/spark/blob/master/LICENSE | 	https://github.com/apache/spark/tree/master/python/pyspark |
-| chedispy | [Proprietary License - Apollomed Medical Holdings Inc.](https://www.apollomed.net/) | | |
+| `chedispy` | [Proprietary License - Apollomed Medical Holdings Inc.](https://www.apollomed.net/) | | |
 
 
-## Project support 
+## Chedispy Installation
 
-Please note the code in this project is provided for your exploration only, and are not formally supported by Databricks with Service Level Agreements (SLAs). They are provided AS-IS and we do not make any guarantees of any kind. Please do not submit a support ticket relating to any issues arising from the use of these projects. The source in this project is provided subject to the Databricks [License](./LICENSE). All included or referenced third party libraries are subject to the licenses set forth below.
-
-Any issues discovered through the use of this project should be filed as GitHub Issues on the Repo. They will be reviewed as time permits, but there are no formal SLAs for support. 
+`chedispy` is ApolloMed's certified,
